backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
boto3==1.40.17
botocore==1.40.17
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.116.1
fonttools==4.59.1
h11==0.16.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
